requirements.txt
aiohttp==3.7.4.post0
asgiref==3.5.0
async-timeout==3.0.1
attrs==21.4.0
backports.zoneinfo==0.2.1
chardet==4.0.0
discord==1.7.3
discord-py-slash-command==3.0.3
discord.py==1.7.3
idna==3.3
multidict==6.0.2
Pillow==9.1.0
psycopg==3.0.12
python-dotenv==0.20.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.1.1
yarl==1.7.2
